boiler plate for SWA 85 86 113
