+++ src/lib/Scenes/Consignments/Screens/SubmitArtworkOverview/ArtworkDetails/ArtworkDetails.tests.tsx
@@ -1,0 +1,1 @@
+// this is a test file
